Extract print_golem_message() and api_request() from testing scripts into a separate module
