+++ concent_api/api-integration-test.py
@@ -5,7 +5,6 @@ import sys
 import datetime
 import random
 import time
-import http.client
 from base64                 import b64encode
 
 from golem_messages         import dump
@@ -13,10 +12,11 @@ from golem_messages         import load
 from golem_messages.message import AckReportComputedTask
 from golem_messages.message import ComputeTaskDef
 from golem_messages.message import ForceReportComputedTask
-from golem_messages.message import Message
 from golem_messages.message import TaskToCompute
 
 from utils.testing_helpers  import generate_ecc_key_pair
+
+from api_testing_helpers            import api_request
 
 import requests
 
@@ -24,80 +24,6 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "concent_api.settings")
 
 (PROVIDER_PRIVATE_KEY,  PROVIDER_PUBLIC_KEY)  = generate_ecc_key_pair()
 (REQUESTOR_PRIVATE_KEY, REQUESTOR_PUBLIC_KEY) = generate_ecc_key_pair()
-
-
-def print_golem_message(message, private_key, public_key, indent = 4):
-    assert isinstance(message, Message)
-    HEADER_FIELDS  = ['timestamp', 'encrypted', 'sig']
-    PRIVATE_FIELDS = {'_payload', '_raw'}
-    assert 'type' not in message.__slots__
-    fields = ['type'] + HEADER_FIELDS + sorted(set(message.__slots__) - set(HEADER_FIELDS) - PRIVATE_FIELDS)
-    values = [
-        type(message).__name__ if field == 'type' else
-        getattr(message, field)
-        for field in fields
-    ]
-
-    for field, value in zip(fields, values):
-        if isinstance(value, bytes):
-            try:
-                nested_message = load(
-                    value,
-                    private_key,
-                    public_key,
-                    check_time=False,
-                )
-            except InvalidSignature as exception:
-                print("Failed to decode a Golem Message.")
-            if nested_message is None:
-                print('{}{:30} = <BINARY DATA>'.format(' ' * indent, field))
-            else:
-                print('{}{:30} ='.format(' ' * indent, field))
-                print_golem_message(nested_message, private_key, public_key, indent = indent + 4)
-        else:
-            if isinstance(value, Message):
-                print_golem_message(value, private_key, public_key, indent = indent + 4)
-            else:
-                print('{}{:30} = {}'.format(' ' * indent, field, value))
-
-
-def api_request(host, endpoint, private_key, public_key, data = None, headers = None, ):
-    assert all(value not in ['', None] for value in [endpoint, host, headers])
-    url = "{}/api/v1/{}/".format(host, endpoint)
-
-    if data is None:
-        print('RECEIVE ({})'.format(url))
-    else:
-        print('SEND ({})'.format(url))
-        print('MESSAGE:')
-        print_golem_message(data, private_key, public_key)
-
-        data = dump(
-            data,
-            private_key,
-            public_key,
-        )
-
-    if data is None:
-        response = requests.post("{}".format(url), headers = headers)
-    else:
-        response = requests.post("{}".format(url), headers = headers, data = data)
-
-    if len(response.content) != 0:
-        decoded_response = load(
-            response.content,
-            private_key,
-            public_key,
-            check_time = False
-        )
-        print('STATUS: {} {}'.format(response.status_code, http.client.responses[response.status_code]))
-        print('MESSAGE:')
-        print_golem_message(decoded_response, private_key, public_key)
-    else:
-        print('STATUS: {} {}'.format(response.status_code, http.client.responses[response.status_code]))
-        if response.text not in ['', None]:
-            print('RAW RESPONSE: {}'.format(response.text))
-    print()
 
 
 def parse_command_line(command_line):
